Extract out event manager in command parser
Command parser will now only extract out the command type from
the user input and then return the CommandType enum.
